requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
SQLAlchemy==2.0.27
psycopg2==2.9.9
python-dotenv==1.0.1
pyhumps==3.8.0